Revert "bandmates list"
This reverts commit f40f2838ae9bf06990b57730ad379bb71503b1d8.

+++ src/www/updateDBInstruments.php
@@ -1,7 +1,15 @@
 <?php
 session_start();
 
-require 'DBConnect.php';
+$servername = "tooney-lunes";
+$username = "root";
+$password = "";
+$dbname = "tuneyloonsdb";
+
+$conn = new mysqli($servername, $username, $password, $dbname);
+if ($conn->connect_error) {
+	die("Connection failed: " . $conn->connect_error);
+}
 
 $sql = "DELETE FROM user_instruments
 		WHERE user_id = '". $_SESSION['user_id'] ."';";
@@ -12,7 +20,7 @@ if (!empty($_GET)) {
 	foreach ($_GET as $i => $instrName) {
 		$sql = "SELECT instr_id
 				FROM instrument
-				WHERE display_name = '". rawurldecode($instrName) ."';";
+				WHERE instr_name = '". rawurldecode($instrName) ."';";
 		$result = $conn->query($sql);
 
 		while ($row = $result->fetch_assoc()) {
